Updated Evo README to include new FlashAttention version dependencies, installation recommendations, and troubleshooting related to newer versions of NumPy. Renamed semantic mining to semantic design
