+++ semantic_design/sample_reference_sequences/modabc_operon_sequences.fasta
